backend: consolidated gameplay features
