--- backend/src/lib/db/games/games.repository.ts
@@ -1,84 +1,0 @@
-import { type Logger, Result } from "@guillaume-docquier/tools-ts"
-import { eq } from "drizzle-orm"
-import type { AccountId } from "#api/accounts/AccountId.ts"
-import type { GameId } from "#api/games/GameId.ts"
-import type { PlayerId } from "#api/games/PlayerId.ts"
-import { PostgresRepository } from "#lib/db/PostgresRepository.ts"
-import { playersTable, gamesTable } from "#lib/db/schema.ts"
-import { couldNot } from "#lib/errors.ts"
-
-type NewGameRow = typeof gamesTable.$inferInsert
-type GameRow = typeof gamesTable.$inferSelect
-
-export type NewGameModel = NewGameRow
-export type GameModel = GameRow
-
-/**
- * @deprecated To be replaced by better repositories
- */
-export class GamesRepository extends PostgresRepository {
-  private readonly logger: Logger
-
-  /**
-   * @deprecated To be replaced by better repositories
-   */
-  public constructor({ logger, db }: { logger: Logger; db: PostgresRepository["db"] }) {
-    super({ db })
-    this.logger = logger.child({ scope: "games-repository" })
-  }
-
-  /**
-   * @deprecated To be replaced by better repositories
-   */
-  public async updateGame(
-    { gameId }: { gameId: GameId },
-    game: Partial<NewGameModel>,
-    db: PostgresRepository["db"] = this.db,
-  ): Promise<Result<true, string>> {
-    const updateResult = await Result.tryCatch(db.update(gamesTable).set(game).where(eq(gamesTable.id, gameId)))
-
-    if (Result.isFailure(updateResult)) {
-      this.logger.error("Could not update game", { gameId, game, error: updateResult.error })
-      return Result.Failure(couldNot("update game"))
-    }
-
-    return Result.Success(true)
-  }
-
-  /**
-   * @deprecated To be replaced by better repositories
-   */
-  public async endGameWithWinner(
-    { gameId, winnerAccountId }: { gameId: GameId; winnerAccountId: AccountId },
-    db: PostgresRepository["db"] = this.db,
-  ): Promise<Result<true, string>> {
-    const endResult = await Result.tryCatch(async (): Promise<true> => {
-      await db.update(gamesTable).set({ endedAt: new Date(), winnerAccountId }).where(eq(gamesTable.id, gameId))
-
-      return true
-    })
-
-    if (Result.isFailure(endResult)) {
-      this.logger.error("Could not end game with winner", { gameId, winnerAccountId, error: endResult.error })
-      return Result.Failure(couldNot("end game with winner"))
-    }
-
-    return endResult
-  }
-
-  /**
-   * @deprecated To be replaced by better repositories
-   */
-  public async getPlayerIds({ gameId }: { gameId: GameId }, db: PostgresRepository["db"] = this.db): Promise<Result<PlayerId[], string>> {
-    const gamePlayersResult = await Result.tryCatch(
-      db.select({ playerId: playersTable.playerId }).from(playersTable).where(eq(playersTable.gameId, gameId)),
-    )
-
-    if (Result.isFailure(gamePlayersResult)) {
-      this.logger.error("Could not get player ids", { gameId, error: gamePlayersResult.error })
-      return Result.Failure(couldNot("get player ids"))
-    }
-
-    return Result.Success(gamePlayersResult.value.map(({ playerId }) => playerId))
-  }
-}
