+++ frontend/src/features/games/LobbyPage.tsx
@@ -39,7 +39,7 @@ function Game({ game }: { game: ApiTypes.Lobby }): ReactElement {
   const backendApiClient = useBackendApiClient()
   const joinGame = useMutation(backendApiClient.lobbies.join.mutationOptions())
   const leaveGame = useMutation(backendApiClient.lobbies.leave.mutationOptions())
-  const startGame = useMutation(backendApiClient.games.start.mutationOptions())
+  const startGame = useMutation(backendApiClient.gameplay.start.mutationOptions())
 
   return (
     <div className="flex flex-col gap-6">
